Use None for non-existant entry ids, fix 'Next entry' issue & key sorting.
Now uses 'None' as the id for a new entry, until it is saved (at which
point a proper id is generated). Also uses 'None' when asking for
First/Last id for a day with no entries. The next button is not
enabled until you have saved an entry for a particular day. The ids
for each day are now sorted before being returned (dictionary keys are
not guarenteed to be sorted by default).

* lib/tk_main.py
(_SetEntryFormDate): Changed default id to -1 (this is replaced by the
first id for that date). Now expects None where first/last/current id
does not exist in the data file. Fixed 'Next entry' button to only be
enabled when there is at least one entry already.

* lib/tk_main.py
(_NextButtonActivated): renamed newid to nextid for clarity.

* lib/tk_main.py
(_PrevButtonActivated): renamed newid to previd for clarity.

* lib/tk_main.py
(_SaveEntriesToPath): now generates an id for an entry if it does not
already have one. Enables the 'Next entry' button.

* lib/tk_main.py
(_FileRevertMenu): does not convert 'id' to an int as that does not
work when id is None.

* lib/tk_data.py
(get_first_id, get_last_id): Sort the entry keys. Return None if no entries.

* lib/tk_data.py
(get_id_pos, get_prev_id): Sort the entry keys.

* lib/tk_data.py
(get_next_id): Sort the entry keys. Return None if no more entries.