On the 'templates-feature' branch, begin working on issue #27 -- support for entry templates.
This is actually two different days' of work on this issue.  The first
was some changes to the data structure (tk_data.py), untested.  The
second is changes to the UI (tk_main.py), again untested.  There's no
guarantee that these two ideas even line up, because I thought I'd
lost the first patch when I started on the second.  I've since found
that first patch, and its included here.  This commit is happening
because I don't want to again come close to losing this work, even if
ultimately it is deemed unacceptable toward the completion of this
feature.

Version control is good.  If you use it.

* lib/tk_data.py
  (TKEntries.set_template, TKEntries.get_template): New.
  (TKDataParser.TKJ_TAG_TEMPLATE): New.
  (TKDataParser._raise_missing_attr, TKDataParser._raise_unexpected_tag):
    New functions.
  (TKDataParser.startElement): Handle <template>.  Only allow the
    global <author> tag in version 1 of the datafile format.
  (TKDataParser.endElement): Handle </template>.  Convert global
    author into template author.

* lib/tk_main.py
  (TKEventTree.__init__): Add 'Templates' top-level tree item.
  (TKEventTree.PruneAll): Override the default implementation to
    delete entries and templates, but not their top-level containers.
  (TKEventTree._MakeNullData): New helper.
  (TKEventTree.GetDateStack, TKEventTree.EntryChangedListener): Deal
    with extra offset needed due to addition of another level of tree
    hierarchy.
  (ThotKeeper.OnInit): Line re-wrapping, and no longer hide the root
    of the entry tree.